sha2: remove RISC-V hack for unaligned loads
