+++ sha2/src/sha512/riscv_zknh/compact.rs
@@ -2,14 +2,7 @@ use super::{sha512sig0, sha512sig1, sha512sum0, sha512sum1};
 use crate::consts::K64;
 
 #[target_feature(enable = "zknh")]
-pub(in super::super) fn compress(state: &mut [u64; 8], blocks: &[[u8; 128]]) {
-    for block in blocks.iter().map(super::utils::load_block) {
-        compress_block(state, block);
-    }
-}
-
-#[target_feature(enable = "zknh")]
-fn compress_block(state: &mut [u64; 8], mut block: [u64; 16]) {
+pub(super) fn compress_block(state: &mut [u64; 8], mut block: [u64; 16]) {
     let mut s = *state;
 
     for r in 0..80 {
@@ -64,4 +57,35 @@ fn ch(x: u64, y: u64, z: u64) -> u64 {
 #[inline(always)]
 fn maj(x: u64, y: u64, z: u64) -> u64 {
     (x & y) ^ (x & z) ^ (y & z)
+}
+
+/// This function returns `k[R]`, but prevents compiler from inlining the indexed value
+fn opaque_load<const R: usize>(k: &[u64]) -> u64 {
+    assert!(R < k.len());
+    #[cfg(target_arch = "riscv64")]
+    unsafe {
+        let dst;
+        core::arch::asm!(
+            "ld {dst}, 8 * {R}({k})",
+            R = const R,
+            k = in(reg) k.as_ptr(),
+            dst = out(reg) dst,
+            options(pure, readonly, nostack, preserves_flags),
+        );
+        dst
+    }
+    #[cfg(target_arch = "riscv32")]
+    unsafe {
+        let [hi, lo]: [u32; 2];
+        core::arch::asm!(
+            "lw {lo}, 8 * {R}({k})",
+            "lw {hi}, 8 * {R} + 4({k})",
+            R = const R,
+            k = in(reg) k.as_ptr(),
+            lo = out(reg) lo,
+            hi = out(reg) hi,
+            options(pure, readonly, nostack, preserves_flags),
+        );
+        ((hi as u64) << 32) | (lo as u64)
+    }
 }
